HOMIE-toolkit/requirements.txt
numpy==2.2.2
opencv-python-headless
h5py==3.12.1
scipy==1.15.1
rerun-sdk==0.29.0
tqdm==4.67.1
